==> 10_code/30_mortality_data.ipynb <==
## Mortality Data

import pandas as pd
import numpy as np


### Read Data

- Unzip using the unzip filename command in terminal

- Update relative path to file on url below


col_names = ["Year", "358 Cause recoded", "State", "County"]

col_width = [(101, 105), (149, 152), (20, 22), (22, 25)]

# put link to the unzipped file
# You can download `VS07MORT.DUSMCPUB`from 
# https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Datasets/DVS/mortality/mort2007us.zip
data_07 = pd.read_fwf(
    "VS07MORT.DUSMCPUB", colspecs=col_width, names=col_names
)


### Using data from cdc wonder

coc_mapping = pd.read_csv("../00_source_data/COC mapping.csv")


## Drug Deaths

alldeathdata0720 = pd.read_csv("../00_source_data/alldrugdeaths0720.txt", sep="\t")
# alldeathdata0720["Year"].value_counts()


alldata0720merge = pd.merge(
    coc_mapping,
    alldeathdata0720,
    left_on=["STNAME", "FIPS code"],
    right_on=["State", "County Code"],
    how="left",
    indicator=True,
)


alldata0720merge["_merge"].value_counts()


alldeathdata21 = pd.read_csv("../00_source_data/alldrugdeaths21.txt", sep="\t")


alldata21merge = pd.merge(
    coc_mapping,
    alldeathdata21,
    left_on=["STNAME", "FIPS code"],
    right_on=["State", "County Code"],
    how="left",
    indicator=True,
)


alldata21merge["_merge"].value_counts()


alldata = pd.concat([alldata0720merge, alldata21merge])


alldata0721 = alldata[
    [
        "STNAME",
        "CoC Code",
        "Coc",
        "CTYNAME",
        "FIPS code",
        "Year",
        "Deaths",
    ]
]


### Check for Missing drug deaths data

years = [i for i in range(2007, 2022)]
for county in coc_mapping["CTYNAME"].unique():
    tmp_data = alldata0721[alldata0721["CTYNAME"] == county]
    tmp_year = tmp_data["Year"].unique()
    if len(tmp_year) != 15:
        print(f"Data for {county}")
        print(f"Missing years : {list(set(years) - set(tmp_year))}")


# missing Douglas county
dougcounty = alldata0721[(alldata0721["CTYNAME"] == "Douglas County")]
dougcounty["Year"].value_counts()


dougcountyoriginal = alldeathdata0720[(alldeathdata0720["County Code"] == 8035)]
len(dougcountyoriginal["Year"])  # without 2021


### Impute Missing Data

# add 2008 for douglas county as average between 2007 and 2009
doug07deaths = alldata0721[
    (alldata0721["FIPS code"] == 8035) & (alldata0721["Year"] == 2007)
]["Deaths"].values
doug09deaths = alldata0721[
    (alldata0721["FIPS code"] == 8035) & (alldata0721["Year"] == 2009)
]["Deaths"].values
doug08deaths = (doug07deaths + doug09deaths) / 2


doug_08_data = {
    "STNAME": "Colorado",
    "CoC Code": "CO-503",
    "Coc": "Metropolitan Denver",
    "CTYNAME": "Douglas County",
    "FIPS code": 8035,
    "Year": 2008,
    "Deaths": doug08deaths,
}
df_doug_08 = pd.DataFrame(doug_08_data)


alldrugdatafinal = pd.concat([alldata0721, df_doug_08], ignore_index=True)
alldrugdatafinal.tail()


# fixing broomfield
years_available = alldrugdatafinal[alldrugdatafinal["FIPS code"] == 8014]["Year"]
missing_brooomfield_years = list(set(years) - set(years_available))

tmp_dfs = []

for year in missing_brooomfield_years:
    tmp_data = {
        "STNAME": "Colorado",
        "CoC Code": "CO-503",
        "Coc": "Metropolitan Denver",
        "CTYNAME": "Broomfield County",
        "FIPS code": 8014,
        "Year": year,
        "Deaths": np.NaN,
    }
    df = pd.DataFrame(tmp_data, index=[0])
    tmp_dfs.append(df)

broomfield_df = pd.concat(tmp_dfs, ignore_index=True)
broomfield_df


alldrugdatafinal0721 = pd.concat([alldrugdatafinal, broomfield_df], ignore_index=True)
alldrugdatafinal0721["Year"] = alldrugdatafinal0721["Year"].astype(int)
alldrugdatafinal0721.tail(10)


### Combine data

alldrugdatafinal0721 = alldrugdatafinal0721.rename(
    columns={"Deaths": "Drug Deaths"}
).sort_values(by=["CTYNAME", "Year"])
alldrugdatafinal0721


# check length
assert len(alldrugdatafinal0721) == len(coc_mapping) * len(years)

# write data
alldrugdatafinal0721.to_csv("../20_intermediate_files/drug_deaths.csv", index=True)


## Suicide data

allmh0720 = pd.read_csv("../00_source_data/allmhdata0720.txt", sep="\t")
# alldeathdata0720["Year"].value_counts()


allmhdata0720merge = pd.merge(
    coc_mapping,
    allmh0720,
    left_on=["STNAME", "FIPS code"],
    right_on=["State", "County Code"],
    how="left",
    indicator=True,
)


allmhdata0720merge["_merge"].value_counts()


allmh21 = pd.read_csv("../00_source_data/allmhdata21.txt", sep="\t")


allmh21merge = pd.merge(
    coc_mapping,
    allmh21,
    left_on=["STNAME", "FIPS code"],
    right_on=["State", "County Code"],
    how="left",
    indicator=True,
)


allmh21merge["_merge"].value_counts()


allmhdata = pd.concat([allmhdata0720merge, allmh21merge])


allmh0721 = allmhdata[
    [
        "STNAME",
        "CoC Code",
        "Coc",
        "CTYNAME",
        "FIPS code",
        "Year",
        "Deaths",
    ]
]


allmh0721.head()


### Check missing suicide values

years = [i for i in range(2007, 2022)]
for county in coc_mapping["CTYNAME"].unique():
    tmp_data = allmh0721[allmh0721["CTYNAME"] == county]
    tmp_year = tmp_data["Year"].unique()
    if len(tmp_year) != 15:
        print(f"Data for {county}")
        print(f"Missing years : {list(set(years) - set(tmp_year))}")


### Impute missing Broomfield data

# fixing broomfield suicide values
years_available = allmh0721[allmh0721["FIPS code"] == 8014]["Year"]
missing_brooomfield_years = list(set(years) - set(years_available))

tmp_dfs = []

for year in missing_brooomfield_years:
    tmp_data = {
        "STNAME": "Colorado",
        "CoC Code": "CO-503",
        "Coc": "Metropolitan Denver",
        "CTYNAME": "Broomfield County",
        "FIPS code": 8014,
        "Year": year,
        "Deaths": np.NaN,
    }
    df = pd.DataFrame(tmp_data, index=[0])
    tmp_dfs.append(df)

broomfield_mh_df = pd.concat(tmp_dfs, ignore_index=True)
broomfield_mh_df


allmhfinal0721 = pd.concat([allmh0721, broomfield_mh_df], ignore_index=True)
allmhfinal0721["Year"] = allmhfinal0721["Year"].astype(int)
allmhfinal0721.tail(10)


### Combining data

allmhfinal0721 = allmhfinal0721.rename(
    columns={"Deaths": "Suicide Deaths"}
).sort_values(by=["CTYNAME", "Year"])
allmhfinal0721


# check length
assert len(allmhfinal0721) == len(coc_mapping) * len(years)

# write data
allmhfinal0721.to_csv("../20_intermediate_files/suicide_deaths.csv", index=True)
