==> 10_code/40_merger.ipynb <==
## Merging all the files

import pandas as pd
import numpy as np


### Read all intermediate files

drug_deaths = pd.read_csv("../20_intermediate_files/drug_deaths.csv")
suicide_deaths = pd.read_csv("../20_intermediate_files/suicide_deaths.csv")
pop_race = pd.read_csv("../20_intermediate_files/population_race.csv")
hic_pit_pov_unemp = pd.read_csv("../20_intermediate_files/HIC_PIT_POV_UNEMP_2007_2021.csv")


## 🛑Broomfield Na replace when necessary
drug_deaths["Drug Deaths"] = drug_deaths["Drug Deaths"].fillna(0)
suicide_deaths["Suicide Deaths"] = suicide_deaths["Suicide Deaths"].fillna(0)

assert drug_deaths["Drug Deaths"].isna().sum() == 0
assert suicide_deaths["Suicide Deaths"].isna().sum() == 0


### Combine the data

hic_pit_pov_unemp = hic_pit_pov_unemp.sort_values(by=["CoC Number", "year"])
hic_pit_pov_unemp.head(17)


# group by CoC
drug_deaths_grouped = (
    drug_deaths.groupby(["CoC Code", "Coc", "Year"])["Drug Deaths"].sum().reset_index()
)
drug_deaths_grouped.head(17)


suicide_deaths_grouped = (
    suicide_deaths.groupby(["CoC Code", "Coc", "Year"])["Suicide Deaths"]
    .sum()
    .reset_index()
)
suicide_deaths_grouped.head(17)


# group and sum the population df
dfs = []

for col in pop_race.columns[7:]:
    tmp_grp = pop_race.groupby(["CoC Code", "Coc", "Year"])[col].sum().reset_index()
    dfs.append(tmp_grp)


dfs[2]


### Merges

# concat by axis =1

dfs_concat = pd.concat(dfs, axis=1)
dfs_concat


dfs_concat_filtered = dfs_concat[
    [
        "Population",
        "Male",
        "Female",
        "White",
        "Black or African American",
        "American Indian and Alaska Native",
        "Asian",
        "Native Hawaiian and Other Pacific Islander",
        "Two or More Races",
        "Non Hispanic",
        "Hispanic",
    ]
]


dfs_concat_filtered.head()


other_cols = dfs[0][["CoC Code", "Year", "Coc"]]
other_cols.head()


combined_pop_data = pd.concat([dfs_concat_filtered, other_cols], axis=1)
combined_pop_data.head()


# 1 Merge: hic_pit_pov_unemp_merge > population

hic_pit_pov_unemp_merge = pd.merge(
    hic_pit_pov_unemp,
    combined_pop_data,
    left_on=["CoC Number", "year"],
    right_on=["CoC Code", "Year"],
    how="inner",
    indicator=True,
)
hic_pit_pov_unemp_merge.head()


hic_pit_pov_unemp_merge["_merge"].value_counts()



hic_pit_pov_unemp_merge = hic_pit_pov_unemp_merge.drop(
    columns=["_merge", "year", "CoC Code", "Unnamed: 0"]
)


# merge 2: hic_pit_pov_unemp + population > drug deaths
drug_deaths_merge = pd.merge(
    hic_pit_pov_unemp_merge,
    drug_deaths_grouped,
    left_on=["CoC Number", "Year"],
    right_on=["CoC Code", "Year"],
    how="inner",
    indicator=True,
)
drug_deaths_merge.head()


drug_deaths_merge["_merge"].value_counts()


drug_deaths_merge = drug_deaths_merge.drop(columns=["_merge", "CoC Code", "Coc_y"])


# merge 3: hic_pit_pov_unemp + population + drug deaths > suicide deaths

suicide_deaths_merge = pd.merge(
    drug_deaths_merge,
    suicide_deaths_grouped,
    left_on=["CoC Number", "Year"],
    right_on=["CoC Code", "Year"],
    how="inner",
    indicator=True,
)
suicide_deaths_merge.head()


suicide_deaths_merge["_merge"].value_counts()


suicide_deaths_merge = suicide_deaths_merge.drop(
    columns=["_merge", "CoC Code", "Coc_x"]
)


suicide_deaths_merge.head()


final_data = suicide_deaths_merge[
    [
        "CoC Number",
        "Coc",
        "Year",
        "permanent_housing",
        "homeless",
        "poverty",
        "unemployment_rate",
        "Population",
        "Male",
        "Female",
        "White",
        "Black or African American",
        "American Indian and Alaska Native",
        "Asian",
        "Native Hawaiian and Other Pacific Islander",
        "Two or More Races",
        "Non Hispanic",
        "Hispanic",
        "Drug Deaths",
        "Suicide Deaths",
    ]
]


# write data
final_data.to_csv("project_data.csv", index=True)


final_data.head()

# Calculate variable rates per 100,000 
final_data_cp = final_data.copy()

def calculate_rates(col: str,):
    """Calculate rates for each variable per 100,000 people"""
    final_data_cp[col] = (final_data_cp[col]/final_data_cp["Population"])*100000

cols_ = set(final_data.columns)-set(['CoC Number','Coc','Year','unemployment_rate', 'Population'])

for col in cols_:
    calculate_rates(col)

# unemployment rate is calculate by dividing (#unemployed / all people in the labor force) * 100
# Therefore, to calculate unemployment per 100,000 people, we just need to time the unemployment rate by 1,000
final_data_cp['unemployment_rate'] = final_data_cp['unemployment_rate']*1000

final_data_cp.head()

# export data
final_data_cp.to_csv("project_data_rates.csv", index=True)
